# openapi.json
{"openapi":"3.1.0","info":{"title":"Face Recognition API","description":"API for ESP32-CAM face uploads, CompreFace integration, and image search","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/api/upload_camera_face":{"post":{"summary":"Upload Camera Face","description":"Upload a cropped face image from ESP32-CAM.\nRequires X-API-KEY header for authentication.","operationId":"upload_camera_face_api_upload_camera_face_post","parameters":[{"name":"X-API-KEY","in":"header","required":true,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_camera_face_api_upload_camera_face_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/search_by_image":{"post":{"summary":"Search By Image","description":"Search for similar faces by image.\nUses CompreFace search API and filters results by time range and similarity threshold.","operationId":"search_by_image_api_search_by_image_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_search_by_image_api_search_by_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/saved_images":{"get":{"summary":"Get Saved Images","description":"Get paginated list of saved face images with optional filters.\nDefaults to today's images if no dates provided.","operationId":"get_saved_images_api_saved_images_get","parameters":[{"name":"start_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Date"}},{"name":"end_date","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Date"}},{"name":"camera_ids","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Ids"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SavedImagesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cameras":{"get":{"summary":"Get Cameras","description":"Get list of all registered cameras.","operationId":"get_cameras_api_cameras_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/CameraResponse"},"type":"array","title":"Response Get Cameras Api Cameras Get"}}}}}}}},"components":{"schemas":{"Body_search_by_image_api_search_by_image_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"},"start_ts":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Start Ts"},"end_ts":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"End Ts"},"threshold":{"type":"number","title":"Threshold","default":0.7},"limit":{"type":"integer","title":"Limit","default":50}},"type":"object","required":["image"],"title":"Body_search_by_image_api_search_by_image_post"},"Body_upload_camera_face_api_upload_camera_face_post":{"properties":{"camera_id":{"type":"string","title":"Camera Id"},"camera_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Name"},"timestamp":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Timestamp"},"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["camera_id","image"],"title":"Body_upload_camera_face_api_upload_camera_face_post"},"CameraResponse":{"properties":{"id":{"type":"integer","title":"Id"},"code":{"type":"string","title":"Code"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"first_seen":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"First Seen"},"last_seen":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Seen"},"total_images":{"type":"integer","title":"Total Images"}},"type":"object","required":["id","code","total_images"],"title":"CameraResponse","description":"Camera information response."},"CapturedImageResponse":{"properties":{"id":{"type":"integer","title":"Id"},"camera_id":{"type":"string","title":"Camera Id"},"camera_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Name"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"image_url":{"type":"string","title":"Image Url"},"compreface_face_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Compreface Face Id"}},"type":"object","required":["id","camera_id","timestamp","image_url"],"title":"CapturedImageResponse","description":"Captured image response schema."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"database":{"type":"string","title":"Database"},"compreface":{"type":"string","title":"Compreface"},"stats":{"anyOf":[{"type":"object"},{"type":"null"}],"title":"Stats"}},"type":"object","required":["status","database","compreface"],"title":"HealthResponse","description":"Health check response."},"SavedImagesResponse":{"properties":{"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"total_pages":{"type":"integer","title":"Total Pages"},"items":{"items":{"$ref":"#/components/schemas/CapturedImageResponse"},"type":"array","title":"Items"}},"type":"object","required":["total","page","page_size","total_pages","items"],"title":"SavedImagesResponse","description":"Paginated saved images response."},"SearchResponse":{"properties":{"results":{"items":{"$ref":"#/components/schemas/SearchResult"},"type":"array","title":"Results"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["results","total"],"title":"SearchResponse","description":"Search by image response."},"SearchResult":{"properties":{"image_url":{"type":"string","title":"Image Url"},"camera_id":{"type":"string","title":"Camera Id"},"camera_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Name"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"similarity":{"type":"number","title":"Similarity"}},"type":"object","required":["image_url","camera_id","timestamp","similarity"],"title":"SearchResult","description":"Search result item."},"UploadResponse":{"properties":{"id":{"type":"integer","title":"Id"},"image_url":{"type":"string","title":"Image Url"},"camera_id":{"type":"string","title":"Camera Id"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"compreface_face_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Compreface Face Id"}},"type":"object","required":["id","image_url","camera_id","timestamp"],"title":"UploadResponse","description":"Response after uploading a camera face."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}